将用数据库存储替换成用区块链存储，将raft算法应用到区块链共识中，下一步需要做的是修改区块链UTXO数据结构，以满足业务需求

+++ fundraising/src/main/java/com/gdut/fundraising/entities/GiftTblEntity.java
@@ -21,7 +21,7 @@ public class GiftTblEntity {
     /**
      * 募捐的钱数
      */
-    private double giftMoney;
+    private long giftMoney;
     /**
      * 募捐对应项目的id
      */
